Rename ee to hierarchical
