+++ src/ops/hierarchical/composition_config_generator.py
