+++ src/ops/hierarchical/config_generator.py
